requirements.txt
openai
langchain
qdrant_client
jupyter
python-decouple
tiktoken